update location map
